Moved content helper into its own file.
