+++ lib/factories/index.js
@@ -6,7 +6,7 @@ const JSZip = require('jszip');
 const async = require('async');
 const xml2js = require('xml2js');
 
-//let { ExcelHelper } = require('../helpers/excel-helper');
+let { PptxContentHelper } = require('../helpers/pptx-content-helper');
 let { ContentTypeFactory } = require('./content-types');
 let { DocPropsFactory } = require('./doc-props');
 let { PptFactory } = require('./ppt');
@@ -16,46 +16,6 @@ let { Image } = require('../image');
 let { TextBox } = require('../text-box');
 let { Shape } = require('../shape');
 let { Chart } = require('../chart');
-
-// TODO: think of a better name for this...
-class PptxContentHelper {
-    // Given the "content" block from the root (ex: PowerPointFactory.content), this function will pull out every slide and return very basic info on them.
-    // (Right now, it's just the slide layout name used on each slide and the relId for that layout-to-slide relationship.)
-    static extractInfoFromSlides(content) {
-        let slideInformation = {}; // index is slide name
-
-        for (let key in content) {
-            if (key.substr(0, 16) === 'ppt/slides/slide') {
-                let slideName = key.substr(11, key.lastIndexOf('.') - 11);
-                let slideRelsKey = `ppt/slides/_rels/${slideName}.xml.rels`;
-                let slideLayoutRelsNode = content[slideRelsKey]['Relationships']['Relationship'].filter(function(element) {
-                    return element['$']['Type'] === 'http://schemas.openxmlformats.org/officeDocument/2006/relationships/slideLayout';
-                })[0];
-
-                let relId = slideLayoutRelsNode['$'].Id;
-                let target = slideLayoutRelsNode['$'].Target;
-                let layout = target.substr(target.lastIndexOf('/') + 1);
-                layout = layout.substr(0, layout.indexOf('.'));
-
-                let objectInfo = PptxContentHelper.extractSlideObjectInfo(content, slideName);
-
-                slideInformation[slideName] = { layout: { relId: relId, name: layout }, objectCount: objectInfo.objectCount };
-            }
-        }
-
-        return slideInformation;
-    }
-
-    static extractSlideObjectInfo(content, slideName) {
-        let objectInfo = {
-            objectCount: 0,
-        };
-
-        // TODO... Mark: you can add code here...
-
-        return objectInfo;
-    }
-}
 
 class PowerPointFactory {
     constructor(presentation, args) {
